Validate format of issuer in ServiceProvider model
**Why**: This verifies that service providers use a clear and consistent
schema for names. Since this is a validation and not a database
constraint,

Add a hint about issuer format to the service provider form

**Why**: Users creating or updating service providers need to know what
the valid format is. This commit adds a hint to the form that describes
the proper format

Update the issuer hint and value when a user switches between SAML and
OpenID connect on the service provider form.

**Why**: When the user updates between SAML and OpenID connect, they
will want to modify the format of their issuer. This commit modifies
the existing Javascript event listener to handler that step for them.

Compose issuer from department and app substrings

**Why**: We were asking users to enter the issuer string in a valid
format which had some UX concerns. This commit changes the form so that
the issuer is constructed from a combination of a department and app
substrings that the user enters and the protocol they select.

Use formatted issuer for idp in setup rake task

**Why**: The rake task that added the local idp to the database was
using a UUID for the issuer. This was causing a validation error when
the task was run since the issuer was invalid.

Make issuer attribute readonly

**Why**: The issuer attribute functions as a primary key in several
parts of our system. As such it is a challenge to modify the issuer for
a service provider. With this in mind, this commit changes the issuer so
it is invalid. Additionally, it changes the code that builds and
validates the formats of issuers such that legacy service providers with
malformed issuers can be modified without any validation errors.