Allow admins to create users
**Why**: Currently, access to the dashboard is protected with basic
auth. As that's being removed with identitysandbox.gov, we need another
way to manage access. This introduces the ability for admins to create
users by email address. When a user logs in, if their email is present,
their login.gov UUID will be associated and they will be given access.
If there is no user by email, they will be stopped and directed to their
login.gov contact for assistance.