Remove redundant secret_key_base definitions.
Only define secret_key_base in application.yml/Figaro. Reference that
value in secrets.yml so it is used by Rails and Devise.

This simplifies the configuration and makes it much easier to understand
where the values are actually coming from.

Remove production values from application.yml and add a note that the
entire file is replaced wholesale by chef in production.