Set default for IAL input to current value.
All of the form fields in the edit form ought to default to the current
values from the object. That way if a user hits edit, makes no changes,
and hits change, the saved object will be unchanged.

This fixes a bug that would have led to unexpected IAL changes when
users edited service providers.