Allow admins to enable prompt=login for SPs
Add boolean allow_prompt_login attribute to the service_providers table
that defaults to false, matching what the IdP has. Conditionally add
checkbox field for admins on the service provider form to allow editing
of that attribute, and also authorize editing of that attribute at the
controller level. Finally, add display for this attribute in the show
view.

Includes a feature spec for the happy path and unit tests for a new
helper method for the show view.