Remove active session check and timeout logic
I was trying to convert the JS for this timeout modal from jQuery to vanilla JS. I could not figure out where any of it was used so I've just removed it in this commit. After a bit of digging around in the history I'm pretty confident this code has never actually been run.