Show admins the audit trail for added and deleted users
I wanted to reuse some of our existing PaperTrail presentation code, but it didn't quite work because the join table history itself doesn't have enough information.

Instead, this commit includes:

* creating a model that inherits from a Struct to manipulate the user change data into something easier to display
* creating a different table format on the teams show view to display that data
* a new PaperTrail::VersionPolicy pundit clas
* a feature spec testing a simple scenario
* some functionality and tests around the policy class

The main shortcoming here is that there is a scenario where a deleted user's email will be missing from the table. For users who were really deleted entirely from the system, the audit event queries are not cross-referencing the user table paper trail to make sure that we have the user's email from before user deletion. In that scenario, the table on the teams page will have a user ID and the user email will be blank.

This change also adds an n+1 query in `TeamMembershipAuditEvent.from_versions`. I think this performance hit in exchange for code readability is acceptable as we don't anticipate the number of users per team being large enough yet for a significant performance hit.