Refactor help text concerns into their own class
There's more work to do on the help text, so let's make that easier through a refactor.

Sandi Metz says that if you have lots of methods that all begin the same (in this case, `help_text_`), that's usually an object screaming to get out.

In my experience, if you refactor that out into a ~100 line class file, then yes, it was time to do the refactor.

The code was easy to refactor. Thank you to everyone who worked on this before me for writing code that's easy to do maintenance work on.