Ensure SAML options persist when using OIDC; some users rely on this
This is an undocumented feature. Since we have organizations that rely on it, it should have a test case — shout out to AJ for suggesting it should.

This commit also stops the user_id from the service provider from unnecessarily churning. This shouldn't cause any functional problems, since the users should always be on the same team and otherwise we don't care much about the user<->service_provider association. Still, why cause problems if they're easy to prevent