Add a `post_idv_follow_up_url` configuration to service providers
* update identity-validations tag

* fix broken sp controller test

* update identity validations

* Add a `post_idv_follow_up_url` configuration to service providers

In https://github.com/18F/identity-idp/pull/11591 we added a new attribute named `post_idv_follow_up_url` to service providers. This attribute represents a URL that is displayed to to users when they have finished proofing but have not gone back to sign in to a service provider.

This commit adds this new URL to the dashboard so that service provider maintainers can configure this value for their applications.