Do not allow users to edit post-IdV follow-up URL
In a previous commit we added the `post_idv_follow_up_url` to the service provider. We allowed users to modify that field, but determined we actually don't want that until we shore up some functionality in the IDP. This commit removes the ability for users to edit that field.