LG-15393 Allow SPs to update the Post-IdV follow-up URL on the dashboard
In !165 we removed the ability for partners to make changes to post-IdV follow-up while we did some work on the IdP to make that attribute safe to use. The work on the IdP is captured in this pull request: https://github.com/18F/identity-idp/pull/11781

When that pull request is merged and deploy it will be safe to merge and deploy this change to allow partners to edit this value.