LG-15851 Stop writing the post-IdV follow-up URL to the IdP
* LG-15851 Stop writing the post-IdV follow-up URL to the IdP

In !165 we stopped allowing partners to edit the post-Idv follow-up URL on the dashboard. We began manually configuring this value for partners who need it on the IdP.

This commit stops the IdP from receiving this value when a service provider is updated on the dashboard. This fixes a bug that occured when the SP was updated an a null `post_idv_follow_up_url` would be written to the IdP overwriting the existing value.